Do not allow ToggleOutput() and WaitForEdge() to set the GPIO direction.
Previously, ToggleOutput() and WaitForEdge() would set the GPIO
direction to in/out on behalf of the user. The latter is a faulty
approach in that changing the direction of the pin, even if the new
direction matches the old, causes the pin value to change! This update
has either method immediately return if the GPIO pin direction is not
set to the value required expected by method (i.e., you cannot toggle
input pins and you cannot wait on output pins).